workspace switching works
